pyproject.toml
[tool.poetry]
name = "books-service"
version = "0.1.0"
description = "Minimal implementation for a book rent service"
authors = ["Mohammed Mwijaa <mm.mwijaa@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
dj-database-url = "^0.5.0"
django-rest-framework = "^0.1.0"
python-decouple = "^3.4"
gunicorn = "^20.1.0"
boto3 = "^1.17.49"
django-storages = "^1.11.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
requests = "^2.25.1"
flake8 = "^3.9.0"
black = "^20.8b1"
pre-commit = "^2.12.0"
pytest-django = "^4.1.0"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-raq --cov=. --cov-report html --ds=core.settings"
testpaths = [
	"tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
